Use separate code file
